Add array literal support to virtual machine and compiler
The commit introduces the ability to compile and execute array literals in the virtual machine and compiler. It also adds tests to ensure correct functionality. This includes changes to the compiler to generate appropriate instructions for array creation, and modifications to the virtual machine to execute these instructions. With this update, array literals are fully supported and can be used in code efficiently.